adds check for proper env var name, expand env inside double quotes, prints quotes inside quotes
