+++ minienv/minienv_str_utils.c
@@ -1,0 +1,37 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   minienv_str_utils.c                                :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/25 14:52:39 by lalex-ku          #+#    #+#             */
+/*   Updated: 2022/05/25 14:52:57 by lalex-ku         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*name_only(char *key_pair)
+{
+	int	i;
+
+	i = 0;
+	while (key_pair[i] != '=' && key_pair[i])
+		i++;
+	if (!key_pair[i])
+		return (NULL);
+	return (ft_substr(key_pair, 0, i));
+}
+
+char	*value_only(char *key_pair)
+{
+	int	i;
+
+	i = 0;
+	while (key_pair[i] != '=' && key_pair[i])
+		i++;
+	if (!key_pair[i])
+		return (NULL);
+	return (&key_pair[i + 1]);
+}
