+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/25 13:36:50 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/05/25 14:38:41 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/05/25 15:07:00 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,4 +108,7 @@ char	**split_commands(char *input);
 
 // error checker
 int		has_unclosed_quotes(char *str);
+
+// utils 
+void	expand_variables(char **commands, t_env *minienv);
 #endif
