+++ src/minishell.c
@@ -3,68 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/23 14:44:34 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/05/24 18:47:59 by sguilher         ###   ########.fr       */
+/*   Updated: 2022/05/25 14:51:03 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	expand_variable(char **command, t_env *minienv)
-{
-	char	*position;
-	int		name_size;
-	char	*var_name;
-	char	*var_value;
-	char	*aux1;
-	char	*aux2;
-
-	position = ft_strchr(*command, '$');
-	if (position)
-	{
-		name_size = 1;
-		while (position[name_size] && position[name_size] != ' ')
-			name_size++;
-		var_name = ft_substr(position, 1, name_size - 1);
-		var_value = minienv_value(var_name, minienv);
-		position[0] = '\0';
-		if (!var_value)
-			aux1 = ft_strjoin(*command, "");
-		else
-			aux1 = ft_strjoin(*command, var_value);
-		aux2 = ft_strjoin(aux1, &position[name_size]);
-		free(*command);
-		*command = aux2;
-		free(var_name);
-		free(aux1);
-	}
-	if (ft_strchr(*command, '$'))
-		expand_variable(&*command, minienv);
-}
-
-void	expand_variables(char **commands, t_env *minienv)
-{
-	while (*commands)
-	{
-		expand_variable(&*commands, minienv);
-		commands++;
-	}
-}
-
-int	is_empty(char *str)
-{
-	if (!str)
-		return (1);
-	while (*str)
-	{
-		if (*str != ' ')
-			return (0);
-		str++;
-	}
-	return (1);
-}
 
 int	minishell(t_env *minienv)
 {
